# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Playbook API",
    "version": "1.0.0",
...
  },
  "servers": [
    {
      "url": "https://api.playbooklab.com.br",
      "description": "Producao"
    },
    {
      "url": "http://localhost:8787",
      "description": "Ambiente local de desenvolvimento"
    }
  ],
  "security": [
    {
      "apiKey": []
    }
  ],
...
  "paths": {
...
    "/v1/keys/rotate": {
      "post": {
        "tags": [
          "Conta"
        ],
        "summary": "Trocar sua chave",
        "description": "Gera uma chave nova e revoga a chave usada nesta chamada. O valor novo aparece UMA unica vez, na resposta: guarde-o imediatamente.",
        "operationId": "rotateKey",
        "responses": {
          "200": {
            "description": "Chave trocada.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KeyRotatedEnvelope"
                },
                "example": {
                  "ok": true,
                  "data": {
                    "api_key": "lk_live_...nova...",
                    "note": "Guarde agora: este valor nao sera exibido de novo. A chave usada nesta chamada foi revogada."
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "402": {
            "$ref": "#/components/responses/AccountPaused"
          },
          "409": {
            "$ref": "#/components/responses/AccountNotReady"
          }
        }
      }
    },
    "/v1/webhook": {
      "put": {
        "tags": [
          "Conta"
        ],
        "summary": "Configurar webhook de eventos",
        "description": "Registra a URL https do webhook principal da conta (o mais antigo), que recebe `message.received`, `account.disconnected` (com um link de reconexao) e `account.reconnected`. Se ele ja existe, troca a URL e gera um secret novo. Para escolher outros eventos, ou ter mais de um webhook, use o painel. Cada entrega e um POST JSON `{type, data}` assinado: valide `X-Webhook-Signature` (sha256=HMAC-SHA256 do `\"{X-Webhook-Timestamp}.{corpo}\"` com o secret retornado aqui). O secret aparece UMA unica vez.",
        "operationId": "configureWebhook",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "url"
                ],
                "additionalProperties": false,
                "properties": {
                  "url": {
                    "type": "string",
                    "description": "URL https que recebera os eventos."
                  }
                }
              },
              "example": {
                "url": "https://meuapp.com.br/webhooks/linkedapi"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Webhook configurado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookConfiguredEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "402": {
            "$ref": "#/components/responses/AccountPaused"
          },
          "409": {
            "$ref": "#/components/responses/AccountNotReady"
          },
          "413": {
            "$ref": "#/components/responses/PayloadTooLarge"
          }
        }
      },
      "get": {
        "tags": [
          "Conta"
        ],
        "summary": "Ver configuracao do webhook",
        "operationId": "getWebhook",
        "responses": {
          "200": {
            "description": "Configuracao atual (o secret nunca e reexibido).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookStatusEnvelope"
                },
                "example": {
                  "ok": true,
                  "data": {
                    "url": "https://meuapp.com.br/webhooks/linkedapi",
                    "configured": true
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "402": {
            "$ref": "#/components/responses/AccountPaused"
          },
          "409": {
            "$ref": "#/components/responses/AccountNotReady"
          }
        }
      },
      "delete": {
        "tags": [
          "Conta"
        ],
        "summary": "Remover webhook",
        "operationId": "deleteWebhook",
        "responses": {
          "200": {
            "description": "Webhook removido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookStatusEnvelope"
                },
                "example": {
                  "ok": true,
                  "data": {
                    "configured": false
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "402": {
            "$ref": "#/components/responses/AccountPaused"
          },
          "409": {
            "$ref": "#/components/responses/AccountNotReady"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-KEY",
        "description": "Sua chave de API, no formato `lk_live_...`. Enviada em toda requisicao no header `X-API-KEY`."
      }
    },
    "schemas": {
...
      "KeyRotatedEnvelope": {
        "type": "object",
        "required": [
          "ok",
          "data"
        ],
        "properties": {
          "ok": {
            "type": "boolean",
            "const": true
          },
          "data": {
            "type": "object",
            "required": [
              "api_key"
            ],
            "properties": {
              "api_key": {
                "type": "string",
                "description": "Sua chave nova (formato lk_live_...). Exibida somente aqui."
              },
              "note": {
                "type": "string"
              }
            }
          }
        }
      },
      "WebhookConfiguredEnvelope": {
        "type": "object",
        "required": [
          "ok",
          "data"
        ],
        "properties": {
          "ok": {
            "type": "boolean",
            "const": true
          },
          "data": {
            "type": "object",
            "required": [
              "url",
              "secret"
            ],
            "properties": {
              "url": {
                "type": "string"
              },
              "secret": {
                "type": "string",
                "description": "Secret de assinatura (formato lk_whsec_...). Exibido somente aqui; valida o header X-Webhook-Signature de cada evento."
              },
              "note": {
                "type": "string"
              }
            }
          }
        }
      },
      "WebhookStatusEnvelope": {
        "type": "object",
        "required": [
          "ok",
          "data"
        ],
        "properties": {
          "ok": {
            "type": "boolean",
            "const": true
          },
          "data": {
            "type": "object",
            "required": [
              "configured"
            ],
            "properties": {
              "url": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "configured": {
                "type": "boolean"
              }
            }
          }
        }
      },
      "ErrorEnvelope": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "Codigo do erro, legivel por maquina."
          },
          "upstream_status": {
            "type": "integer",
            "description": "Presente em erros 502: o status HTTP recebido do LinkedIn."
          },
          "action": {
            "type": "string",
            "description": "Presente em erros 429: a acao cujo limite foi atingido."
          },
          "limit": {
            "type": "integer",
            "description": "Presente em erros 429: o limite diario da acao."
          },
          "retry_after": {
            "type": "integer",
            "description": "Presente em erros 429: segundos ate a proxima janela (igual ao header Retry-After)."
          },
          "reason": {
            "type": "string",
            "description": "Motivo legivel por maquina. Em 429: `too_many_attempts` quando o teto de tentativas do dia (10x o limite) foi atingido. Em 400 e 502: o codigo de erro informado pelo LinkedIn, quando houver."
          }
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Requisicao invalida: campo obrigatorio ausente, formato errado, JSON malformado, cursor invalido, ou filtro recusado pelo LinkedIn (`invalid_request`, com `reason`).",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            },
            "examples": {
              "missing_field": {
                "value": {
                  "error": "missing_chat_id"
                }
              },
              "invalid_json": {
                "value": {
                  "error": "invalid_json"
                }
              },
              "invalid_cursor": {
                "value": {
                  "error": "invalid_cursor"
                }
              },
              "invalid_request": {
                "value": {
                  "error": "invalid_request",
                  "reason": "invalid_parameters"
                }
              },
              "unknown_filter": {
                "value": {
                  "error": "unknown_filter",
                  "filter": "hiring_project"
                }
              }
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Chave ausente, invalida ou revogada.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            },
            "examples": {
              "missing_api_key": {
                "value": {
                  "error": "missing_api_key"
                }
              },
              "invalid_api_key": {
                "value": {
                  "error": "invalid_api_key"
                }
              }
            }
          }
        }
      },
      "AccountPaused": {
        "description": "Sua conta esta pausada porque o pagamento da assinatura esta em atraso. Ela volta sozinha assim que o Pix for confirmado.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            },
            "example": {
              "error": "account_paused"
            }
          }
        }
      },
      "AccountNotReady": {
        "description": "Sua chave e valida, mas nao ha LinkedIn ativo na conta: ou a sessao caiu (`account_disconnected`, reconecte pelo painel) ou nenhum LinkedIn foi conectado ainda (`linkedin_not_connected`).",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            },
            "examples": {
              "account_disconnected": {
                "value": {
                  "error": "account_disconnected"
                }
              },
              "linkedin_not_connected": {
                "value": {
                  "error": "linkedin_not_connected"
                }
              }
            }
          }
        }
      },
...
      "PayloadTooLarge": {
        "description": "Corpo grande demais: anexos acima de 15 MB somados (`attachments_too_large`) ou JSON acima de 1 MB (`body_too_large`).",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            },
            "example": {
              "error": "attachments_too_large",
              "max_bytes": 15728640
            }
          }
        }
      }
    }
  },
...
}